# server/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: EduCenterPro API
  description: 'Tài liệu API đầy đủ cho EduCenterPro — nền tảng quản lý trung tâm ngoại ngữ.


    Gồm: Auth, Học viên, Giáo viên, Lớp học, Buổi học & Điểm danh, Học phí & Công nợ, Bài tập & Quiz, Sổ điểm, Phòng học,
    Lương, Học thử, Leads CRM, Referral, Reviews, Zalo OA, Cổng phụ huynh, Cổng giáo viên, Public landing.


    Tất cả endpoint staff đều yêu cầu JWT (đăng nhập admin/staff/teacher).

    Base path chính: /api/v1 (alias legacy /api vẫn hoạt động nhưng deprecated).'
  version: 1.0.0
security:
  - bearerAuth: []
servers:
  - url: http://localhost:4000
    description: Dev server
  - url: /
    description: Production (cùng origin)
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Đăng nhập POST /api/auth/login để lấy token
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: Không tìm thấy bài tập
        code:
          type: string
          example: NOT_FOUND
...
paths:
...
  /api/v1/invoices:
    get:
      tags:
        - Invoices
      summary: Danh sách hóa đơn học phí
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum:
              - unpaid
              - partial
              - paid
              - overdue
        - in: query
          name: student_id
          schema:
            type: integer
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: OK
    post:
      tags:
        - Invoices
      summary: Tạo hóa đơn học phí
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - student_id
                - amount
              properties:
                student_id:
                  type: integer
                amount:
                  type: number
                  example: 2500000
                due_date:
                  type: string
                  example: '2026-10-31'
                class_id:
                  type: integer
                  nullable: true
                note:
                  type: string
      responses:
        '201':
          description: Đã tạo
...
  /api/v1/invoices/{id}:
    get:
      tags:
        - Invoices
      summary: Chi tiết hóa đơn
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
    put:
      tags:
        - Invoices
      summary: Cập nhật hóa đơn
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
    delete:
      tags:
        - Invoices
      summary: Xóa hóa đơn
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
  /api/v1/invoices/{id}/payments:
    post:
      tags:
        - Invoices
      summary: Ghi nhận thanh toán
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - amount
              properties:
                amount:
                  type: number
                  example: 2500000
                method:
                  type: string
                  enum:
                    - cash
                    - bank_transfer
                    - vnpay
                  example: cash
                note:
                  type: string
      responses:
        '201':
          description: Đã ghi nhận
  /api/v1/invoices/{id}/refund:
    post:
      tags:
        - Invoices
      summary: Hoàn tiền cho hóa đơn (quyền payments.refund)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - amount
              properties:
                amount:
                  type: integer
                  description: Số tiền hoàn (không vượt số đã thu thực)
                reason:
                  type: string
                  description: Lý do hoàn tiền
      responses:
        '200':
          description: OK — ghi nhận payment âm method='refund', tính lại trạng thái hóa đơn
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  refunded:
                    type: integer
                  status:
                    type: string
        '400':
          description: Số tiền không hợp lệ / vượt số đã thu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/invoices/{id}/apply-credit:
    post:
      tags:
        - Invoices
      summary: Trừ credit giới thiệu vào hóa đơn
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
...
  /api/v1/leaves:
    get:
      tags:
        - Leaves
      summary: Danh sách đơn xin nghỉ
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum:
              - pending
              - approved
              - rejected
      responses:
        '200':
          description: OK
  /api/v1/leaves/{id}/approve:
    post:
      tags:
        - Leaves
      summary: Duyệt đơn nghỉ
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
  /api/v1/leaves/{id}/reject:
    post:
      tags:
        - Leaves
      summary: Từ chối đơn nghỉ
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
...
  /api/v1/payroll:
    get:
      tags:
        - Payroll
      summary: Bảng lương giáo viên (tự tính từ điểm danh; tháng đã chốt trả bảng chụp lúc chốt)
      parameters:
        - in: query
          name: month
          schema:
            type: string
          description: YYYY-MM, mặc định tháng hiện tại
      responses:
        '200':
          description: OK
  /api/v1/payroll/rules:
    put:
      tags:
        - Payroll
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      summary: Cập nhật quy tắc tính lương
      responses:
        '200':
          description: OK
...
  /api/v1/centers:
    get:
      tags:
        - Centers
      summary: Danh sách trung tâm (superadmin)
      responses:
        '200':
          description: OK
    post:
      tags:
        - Centers
      summary: Tạo trung tâm mới (superadmin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - admin_username
                - admin_password
              properties:
                name:
                  type: string
                subdomain:
                  type: string
                phone:
                  type: string
                address:
                  type: string
                plan:
                  type: string
                  enum: [basic, standard, premium]
                admin_username:
                  type: string
                admin_password:
                  type: string
                  description: Tối thiểu 8 ký tự
      responses:
        '201':
          description: Đã tạo
  /api/v1/centers/{id}:
    put:
      tags:
        - Centers
      summary: Cập nhật trung tâm / gói dịch vụ
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
...
  /api/v1/parent/leaves:
    get:
      tags:
        - ParentPortal
      summary: Đơn xin nghỉ của con
      responses:
        '200':
          description: OK
    post:
      tags:
        - ParentPortal
      summary: Gửi đơn xin nghỉ
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - student_id
                - from_date
                - to_date
              properties:
                student_id:
                  type: integer
                class_id:
                  type: integer
                  nullable: true
                from_date:
                  type: string
                  example: '2026-10-20'
                to_date:
                  type: string
                  example: '2026-10-22'
                reason:
                  type: string
      responses:
        '201':
          description: Đã gửi
...